+++ packages/slate-plugins/src/common/plugins/inline-void/withInlineVoid.ts
@@ -1,4 +1,4 @@
-import { SlatePlugin } from '@udecode/core';
+import { SlatePlugin } from '@udecode/slate-plugins-core';
 import { castArray } from 'lodash';
 import { Editor } from 'slate';