+++ packages/slate-plugins/src/decorators/preview/PreviewPlugin.ts
@@ -1,4 +1,4 @@
-import { SlatePlugin } from '@udecode/core';
+import { SlatePlugin } from '@udecode/slate-plugins-core';
 import { decoratePreview } from './decoratePreview';
 import { renderLeafPreview } from './renderLeafPreview';