+++ packages/slate-plugins/src/marks/italic/ItalicPlugin.ts
@@ -1,4 +1,4 @@
-import { SlatePlugin } from '@udecode/core';
+import { SlatePlugin } from '@udecode/slate-plugins-core';
 import { onKeyDownMark } from '../../common/utils/onKeyDownMark';
 import { deserializeItalic } from './deserializeItalic';
 import { renderLeafItalic } from './renderLeafItalic';