+++ packages/slate-plugins/src/marks/subscript/SubscriptPlugin.ts
@@ -1,4 +1,4 @@
-import { SlatePlugin } from '@udecode/core';
+import { SlatePlugin } from '@udecode/slate-plugins-core';
 import { onKeyDownMark } from '../../common/utils/onKeyDownMark';
 import { MARK_SUPERSCRIPT } from '../superscript/types';
 import { deserializeSubscript } from './deserializeSubscript';